Fix FH binding and audio bugs found in review; include pending local work
FH bindings (fhapi):
- quad setViewport read args[0-3] instead of args[1-4], corrupting the
  quad with the object pointer reinterpreted as a double
- image setPixel/getPixel validated wrong arg ranges and read args[6]
  without any n_args check; setPixel now also accepts Image objects
  (the else-branch was unreachable before)
- window setVsync passed &args[0] (a pointer, always truthy) instead of
  the boolean value; setIcon/getViewport gained n_args guards
- getShader/getFont return null instead of wrapping NULL in a c_obj

Audio:
- source/streamsource free(): detach buffer BEFORE alDeleteSources
  (was using the source after deletion)
- streamsource stop() set the state to stopped before checking for
  playing, so alSourceStop was dead code
- unchecked realloc of the playing-stream list could lose the pointer
- loadStream leaked decoderData on failure and only caught
  VORBIS_file_open_failure; any decoder error now fails cleanly
- vorbis loadStream dereferenced NULL when the file could not be opened

System / main loop:
- clipboard get/set now own their buffer (no leak, no borrowed pointer)
- love_focus/love_quit are optional: skipped via fh_function_exists()
  instead of polluting the program error state every frame

Also includes pending local changes to graphics (batch, fonts, gltools),
microtar and the FH/Lua examples that predate this session.

+++ src/include/batch.h
@@ -1,7 +1,7 @@
 /*
 #   clove
 #
-#   Copyright (C) 2016-2020 Muresan Vlad
+#   Copyright (C) 2016-2025 Muresan Vlad
 #
 #   This project is free software; you can redistribute it and/or modify it
 #   under the terms of the MIT license. See LICENSE.md for details.
@@ -40,7 +40,8 @@ typedef struct {
 
 void graphics_batch_init(void);
 
-void graphics_Batch_new(graphics_Batch *batch, graphics_Image const *texture, int maxSize, graphics_BatchUsage usage);
+void graphics_Batch_new(graphics_Batch *batch, graphics_Image const *texture, size_t maxSize,
+                        graphics_BatchUsage usage);
 
 int graphics_Batch_add(graphics_Batch *batch, graphics_Quad const *q, float x, float y, float r, float sx, float sy,
                        float ox, float oy, float kx, float ky);
